Fix scrolling inside fixed elements
I have a fixed sidebar in my project. When i call `element.scrollIntoView()` for an element inside that sidebar and there is no scroll in it, the body gets scrolled with the same amount which is definitely unintended behavior. 
This fix has worked for me, though I'm not 100% it will not break something else. Hope you guys can help me with this.