Only attempt to do work in animation frame if there is still scrolling to do
Rather than registring more scroll stepping only to cancel later, delay
registering a new animation frame call until we know if we need it.

This fixes a bug in our application where some other library is replaced the
requestAnimationFrame function to not return a frameId and thus made the
request un-cancelable.  It also however is a simpler way to accomplish the same
behavior.

Signed-off-by: Jonathan Barnes <jbarnes@pivotal.io>