Improve reliability of findScrollableParent
Improves reliability of findScrollableParent by accounting differently for horizontal and vertical overflows. This fixes a bug I encountered where a div was overflowing horizontally causing the whole window to be horizontally scrolled, even though overflow-x: visible; was applied. The commit adds separate checks for overflow-x and overlfow-y. The cases included in index.html seem to be working just fine.

Signed-off-by: Emanuele <feliziani.emanuele@gmail.com>