Improve reliability of findScrollableParent
Improves reliability of findScrollableParent by accounting separately for horizontal and vertical overflows and also by checking for both `scroll` and `auto` values of the overflows, rather than just for `visible`. This fixes a bug that was unnecessarily scrolling items that had `overflow: hidden`. The cases included in index.html seem to be working just fine.

NOTE: improves the previous solution as suggested by one of the collaborators. Also, this commit does NOT include the built file, only the source.
Signed-off-by: Emanuele <feliziani.emanuele@gmail.com>