Comply with TypeScript's convention for object and void
In TypeScript, we avoid the usage of `Object`, favoring `object` instead. Also, we use `void` as the return type of functions that return `undefined`, this indicates that we don't really care that it outputs `undefined`, it is `void`, which means that it's nothing and the return of it should be ignored.