Improve property access type safety
By casting `'overflowX'|'overflowY'` to the `overflowProperty`, we tell the compiler (or type checker) which property we'll be accesing from the `CSSStyleDeclaration`. That's useful for IDEs such as VSCode, that uses the TypeScript parser to analyze the code. It can also be useful if we decide to add a type checker to the codebase later on.